requirements.txt
streamlit>=1.28.0
torch>=1.13.0
torchvision>=0.14.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
gdown>=4.6.0
pandas>=1.5.0
tqdm>=4.64.0
scikit-learn>=1.1.0
